feat: delete p2plog
