+++ config/p2pconfig/p2p_utils.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/libsv/bitcoin-hc/internal/chaincfg"
 	"github.com/libsv/bitcoin-hc/internal/chaincfg/chainhash"
-	"github.com/libsv/bitcoin-hc/transports/p2p/p2plog"
 )
 
 type logWriter struct{}
@@ -23,12 +22,6 @@ func (logWriter) Write(p []byte) (n int, err error) {
 		return len(p), err
 	}
 	return len(p), nil
-}
-// UseDefaultP2PLogger returns default p2p logger.
-func UseDefaultP2PLogger() p2plog.Logger {
-	backendLog := p2plog.NewBackend(logWriter{})
-	logger := backendLog.Logger("HEADERS")
-	return logger
 }
 
 // normalizeAddress returns addr with the passed default port appended if
